refactoring code...
