+++ src/aeos/owl_aeos_types.ml
@@ -39,9 +39,11 @@ type tuner =
   | Sin of Sin.t
   | Cos of Cos.t
   | Add of Add.t
-  | Sum of Sum.t *)
+  | Sum of Sum.t
 
 open Bigarray
 
 type fun_map1 = (float, float32_elt) Owl_core_types.owl_arr_op09
 type fun_map2 = (float, float32_elt) Owl_core_types.owl_arr_op03
+
+*)
